+++ changelog.md
@@ -18,7 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - create management commands
     - create_sample: to create simple sample data
     - easy_setup: auto migrate and create simple sample data
-- add api endpoint for Snack and Machine Model
+- add api endpoint for Snack and Machine model
+- about section in readme file
+- add name field to Machine model
 
 
 ## 2023-01-14